requirements.txt
# 核心框架
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0

# 数据库 - 使用 PyMySQL（Python 3.12 兼容）
PyMySQL==1.1.0

# Redis - 必须升级到 5.0.1+，但推荐最新版
redis==5.0.8

# 缓存
Flask-Caching==2.3.0

# ✅ Celery 全家桶 - 升级到 Python 3.12 兼容版本
celery==5.3.6
kombu==5.3.6
vine==5.1.0
billiard==4.2.0

# ✅ Windows 下运行 Celery 需要
eventlet==0.37.0

# ✅ 监控工具
flower==2.0.1